% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{CommonGames}
\alias{CommonGames}
\title{CommonGames}
\usage{
CommonGames(Team1, Team2, Team3 = NULL, Team4 = NULL, data,
  SimScores = NULL)
}
\arguments{
\item{Team1}{the main team whose statistic is being evaluated by the function}

\item{Team2}{a team that \code{Team1} is being evaluated against}

\item{Team3}{a team that \code{Team1} is being evaluated against (default is \code{NULL})}

\item{Team4}{a team that \code{Team1} is being evaluated against (default is \code{NULL})}

\item{SimScores}{the simulated scores, if a simulation was performed}

\item{Data}{the data frame from which Scores are gathered (default is Week 17 \code{WeeklyUpdate})}
}
\value{
\code{CommonGames} a win-loss record in games the other teams played as well.
\code{CommonGamesPts} the total net points in games the other teams played as well.
}
\description{
These are a pair of functions that will take between 2-4 teams, see which opponents they have in common, and determine
either the teams win-loss record against those teams (\code{CommonGames}) or the net points scored in those games (\code{CommonGamesPts}).
If the total games played against common opponents between all evaluated teams is less than 4, the output will always be 0.
}
\examples{
CommonGames("Patriots","Dolphins")
CommonGames("Patriots","Dolphins","Bills")
CommonGames("Patriots","Dolphins","Bills","Jets")
CommonGamesPts("Patriots","Dolphins")
CommonGamesPts("Patriots","Dolphins","Bills")
CommonGamesPts("Patriots","Dolphins","Bills","Jets")
}
