% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{Conference}
\alias{Conference}
\alias{TwoTieConf}
\alias{ThreeTieConf}
\alias{FourTieConf}
\alias{FinalRank}
\title{Conference Tiebreakers}
\usage{
TwoTieConf(Team1, Team2, scores = WeeklyUpdate())

ThreeTieConf(Team1, Team2, Team3, scores = WeeklyUpdate())

FourTieConf(Team1, Team2, Team3, Team4, scores = WeeklyUpdate())

FinalRank(scores = WeeklyUpdate())
}
\arguments{
\item{scores}{the data frame of scores used for calculating \code{Head-to-Head} and \code{CommonGames}}
}
\value{
A vector of the tiebreaking results corresponding to each team in the function input.
}
\description{
These functions will break any ties between teams with the same win-loss record with in different divisions, but in the same conference
}
\details{
The order of statistics for
which ties are broken are as follows: Apply division tiebreaker to eliminate all but the highest ranked team in each division,
Head-to-Head Record, win-loss record against teams within the conference, win-loss record against teams common opponents (\code{CommonGames}),
strength of victory (win-loss record against all teams defeated), strength of schedule (win-loss record against all teams played),
combined rank for points scored and points allowed among conference teams, combined rank for points scored and points allowed among all teams,
net points against conference opponents, net points overall.

Seedings are as follows: the four division winners/leaders automatically get ranked 1-4 based on record and tiebreakers, while the remaining 12
teams will be seeded based on record and tiebreakers.
}
\examples{
TwoTieConf("Vikings","Giants")
ThreeTieDiv("Vikings","Giants","49ers")
FourTieDiv("Vikings","Giants","49ers","Falcons")
FinalDivRank()
}
