% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{Division}
\alias{Division}
\alias{TwoTieDiv}
\alias{ThreeTieDiv}
\alias{FourTieDiv}
\alias{FinalDivRank}
\title{Division Tiebreakers}
\usage{
TwoTieDiv(Team1, Team2, scores, SimScores = NULL)

ThreeTieDiv(Team1, Team2, Team3, scores, SimScores = NULL)

FourTieDiv(Team1, Team2, Team3, Team4, scores, SimScores = NULL)

FinalDivRank(scores, SimScores = NULL)
}
\arguments{
\item{scores}{the data frame of scores used for calculating \code{Head-to-Head}, \code{CommonGames}, and \code{CommonGamesPts}}

\item{SimScores}{the simulated scores, if a simulation was performed}
}
\value{
A vector of the tiebreaking results corresponding to each team in the function input.
}
\description{
These functions will break any ties between teams with the same win-loss record within the same division.
}
\details{
The order of statistics for
which ties are broken are as follows: Head-to-Head Record, win-loss record against teams within the division, win-loss record against teams common
opponents (\code{CommonGames}), win-loss record against teams within the conference, strength of victory (win-loss record against all teams defeated)
, strength of schedule (win-loss record against all teams played), combined rank for points scored and points allowed among conference teams,
combined rank for points scored and points allowed among all teams, net points against common opponents (\code{CommonGamesPts}), net points overall.
}
\examples{
TwoTieDiv("Vikings","Packers")
ThreeTieDiv("Vikings","Packers","Bears")
FourTieDiv("Vikings","Packers","Bears","Lions")
FinalDivRank()
}
