% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{HeadtoHead}
\alias{HeadtoHead}
\title{Head-to-Head}
\usage{
HeadtoHead(Team1, Team2, Team3 = NULL, Team4 = NULL,
  data = WeeklyUpdate())
}
\arguments{
\item{Team1}{the main team whose game results is being evaluated by the function}

\item{Team2}{a team that \code{Team1} is being evaluated against}

\item{Team3}{a team that \code{Team1} is being evaluated against (default is \code{NULL})}

\item{Team4}{a team that \code{Team1} is being evaluated against (default is \code{NULL})}

\item{Data}{the data frame from which Scores are gathered (default is Week 17 \code{WeeklyUpdate})}
}
\value{
\code{Wins} - \code{Losses} in all games for \code{Team1} against the other teams specified
}
\description{
This function determines the record of one team against up to three other teams in the National Football League
}
\examples{
HeadtoHead("Vikings","Packers")
HeadtoHead("Vikings","Packers","Bears")
HeadtoHead("Vikings","Packers","Bears","Lions")
}
