% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\docType{data}
\name{NFLTeams}
\alias{NFLTeams}
\title{NFLTeams}
\format{A dataset containing 32 rows and 14 variables
\describe{
 \item{Team}{Identifier for the team in the NFL}
 \item{Conference}{Identifier for the team's conference}
 \item{Division}{Identifier for the team's division}
 \item{Wins}{Total Games Won}
 \item{Losses}{Total Games Lost}
 \item{Ties}{Total Games Tied}
 \item{ConfWins}{Total Games Won against teams in their own conference}
 \item{ConfLosses}{Total Games Lost against teams in their own conference}
 \item{ConfTies}{Total Games Tied against teams in their own conference}
 \item{DivWins}{Total Games Won against teams in their own division}
 \item{DivLosses}{Total Games Lost against teams in their own division}
 \item{DivTies}{Total Games Tied against teams in their own division}
 \item{SOV}{Combined Win-Loss Record in games in which Team won}
 \item{SOS}{Combined Win-Loss Record in games played by Team}
}}
\source{
\url{"http://www.dummies.com/sports/football/the-national-football-league-conferences/"}

\url{"http://www.nfl.com/schedules/2016"}
}
\usage{
NFLTeams
}
\description{
Tiebreaking Data for the 32 teams in the National Football League
The dataset containing necessary information for tiebreaking procedures.  The statistics can be updated using the UpdateData function
}
\keyword{datasets}
