% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{SeedPlot}
\alias{SeedPlot}
\title{SeedPlot}
\usage{
SeedPlot(TeamX = "Bills", Sim = NFLSim(data = WeeklyUpdate()))
}
\arguments{
\item{TeamX}{The specific team wanting to be analyzed in the graph}

\item{Sim}{The simulation data set used from the \code{NFLSim} function}
}
\value{
A plotly bar graph showing the frequency of times a teams obtains a specific seed in the conference based on the simulation settings.
The graph is a plotly, meaning you can hover over a specific part of the graph to show the specific seed and the frequency that goes along with it.
}
\description{
This is a function that will take an output from the \code{NFLSim} function and create a plotly of the seeding outcomes from the simulation
}
\examples{
NFLSim()
NFLSim(sims=1000)
}
