% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{Sim}
\alias{Sim}
\alias{NFLSim}
\title{NFLSim}
\usage{
NFLSim(Games = NULL, sims = 100, data = WeeklyUpdate())
}
\arguments{
\item{Games}{The outcomes of any specified games yet to be played (default is \code{NULL}).  Any string of the format (AwayTeam AwayScore HomeTeam
HomeScore) separated by any character not a letter or number will be sufficient}

\item{sims}{Number of simulations requested to be run (default is 100)}

\item{data}{The data frame of scores to be used for simulating scores for games yet to be played, that are not specified by the Games input}
}
\value{
A data frame of each team's playoff seed based on the simulation results
}
\description{
This is a function that will take specified outcomes of games yet to be played, and simulate multiple seedings based on those
specifications.
}
\details{
All simulations will have the scores specified by the Games input.  Any other teams needing scores will be specified in the following way: the mean
and standard deviation from all previous games will be taken for each team, and a score will be generated from a normal distribution with the specified
mean and standard deviation.  Precautions are taken so that scores are a whole number, scores cannot be below zero, and scores cannot be one (a score of
one is impossible in American Football)
}
\examples{
NFLSim()
NFLSim(sims=1000)
NFLSim(Games="Cardinals13Rams17",sims=10)
}
