% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{UpdateTeams}
\alias{UpdateTeams}
\title{UpdateTeams}
\usage{
UpdateTeams(data = WeeklyUpdate())
}
\arguments{
\item{data}{the WeeklyUpdate output that will be used to aggregate the data}
}
\value{
A dataset of the same dimensions and format as NFLTeams
}
\description{
This is a function that will Update all tiebreaking stats from the WeeklyUpdate output and using NFLTeams as a template
}
\examples{
UpdateTeams()
UpdateTeams(WeeklyUpdate(16))
}
