% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{WeeklyGames}
\alias{WeeklyGames}
\title{WeeklyGames}
\format{A dataset containing a varying number of rows and 5 variables
\describe{
 \item{Date}{Day the Game was contested}
 \item{AwayTeam}{The road team in the game}
 \item{AwayScore}{Automatically \code{NA}}
 \item{HomeTeam}{The home team in the game}
 \item{HomeScore}{Automatically \code{NA}}
}}
\source{
\url{"http://www.nfl.com/schedules/2016"}
}
\usage{
WeeklyGames(Week = 1)
}
\arguments{
\item{Week}{the identifier for the week of games you want (Default is 1)}
}
\value{
A dataset of games according to the format description
}
\description{
This is a function that will gets games from weeks that you do not want to use the scores, or if the games have not yet been played that week
}
\examples{
WeeklyGames()
WeeklyGames(Week=2)
}
