% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{WeeklyScores}
\alias{WeeklyScores}
\title{WeeklyScores}
\format{A dataset containing a varying number of rows and 5 variables
\describe{
 \item{Date}{Day the Game was contested}
 \item{AwayTeam}{The road team in the game}
 \item{AwayScore}{The points scored by the road team in the game}
 \item{HomeTeam}{The home team in the game}
 \item{HomeScore}{The points scored by the home team in the game}
}}
\source{
\url{"http://www.nfl.com/schedules/2016"}
}
\usage{
WeeklyScores(Week = 1)
}
\arguments{
\item{Week}{the identifier for the week of scores you want (Default is 1)}
}
\value{
A dataset of Scores of games according to the format description
}
\description{
This is a function that will update the scores from the specified week from NFL.com
}
\examples{
WeeklyScores()
WeeklyScores(Week=2)
}
