% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{WeeklyUpdate}
\alias{WeeklyUpdate}
\title{WeeklyUpdate}
\format{A dataset containing a varying number of rows and 5 variables
\describe{
 \item{Date}{Day the Game was contested}
 \item{AwayTeam}{The road team in the game}
 \item{AwayScore}{The points scored by the road team in the game (\code{NA} if score is not to be recorded)}
 \item{HomeTeam}{The home team in the game}
 \item{HomeScore}{The points scored by the home team in the game (\code{NA} if score is not to be recorded)}
}}
\source{
\url{"http://www.nfl.com/schedules/2016"}
}
\usage{
WeeklyUpdate(WeekID = 17)
}
\arguments{
\item{Week}{the identifier for the latest week of games played you want (Default is 17)}
}
\value{
A dataset of Scores and Games according to the format description
}
\description{
This is a function that will gets all games for the season, with games being played from WeeklyScores and games yet to be played from WeeklyGames
}
\examples{
WeeklyUpdate()
WeeklyUpdate(Week=16)
}
