% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectNFL.R
\name{UpdateTeams}
\alias{UpdateTeams}
\title{UpdateTeams}
\usage{
UpdateTeams(data, SimScores = NULL)
}
\arguments{
\item{SimScores}{A vector of the Games that were simulated}

\item{RealScores}{A vector of the Actual Game Scores}

\item{Games}{A dataframe of the Games for the current season}
}
\value{
A dataset of the same dimensions and format as NFLTeams
}
\description{
This is a function that will Update all tiebreaking stats from the WeeklyUpdate output and using NFLTeams as a template
}
